requirements.txt
Django
djangorestframework
django-filter
Pillow
django-cors-headers
djangorestframework-simplejwt
